Bugfixes, added role debug mode, fixed default configuration
I removed the playerColors dictionary and the netcode that handled the incoming colors from the server. The playerId's stored on the server and client wasn't syncing and it wouldn't allow players to choose their colors. I will fix this with a better way to get colors soon. I added a role debug mode that is toggleable when building that I was using to force the roles of certain players so I could record footage for the release trailer. I also reset the default config values to the reccomended settings. Disabled the shield indicator because it also relied on the faulty colors. It was set back to a simple name color change.