Fixed Engineer venting, Switched type names to Reactor shortcuts, finished medic reports (again)
This is now completely ported to a new framework thingy, so building is more difficult. You have to add the directory to the Among Us modded folder in the PATH, and then you can build if it can find BepInEx and all that stuff.

Also, there is a weird bug where Harmony wont hook the right functions unless the Assembly-CSharp-Deobfuscated.dll is in the plugin folder. Not only is it large, I'm pretty sure it's illegal to distribute. Hm... I'll have to find some way to make a release without this ASAP.

Also, moving to this new framework has let me finish medic reports and make the Joker standing alone at the beggining of the game.