Removed color arrays, reworked icons, fixed end-screen order, fixed Joker not being able to fix Sabotages, changed shielded players visor color
Removed the color arrays and instead loaded the images from a byte array. This wasn't working on the old framework. The icons were remodled. The local player is now standing in front on the winner lineup. The Joker can now fix sabotages, but there is a known bug where "Fix Lights" appears twice in the list. Added a cool indication that a player is shielded by changing their visor color to cyan.